Got all storage tests
